+++ dts/planner/interfaces.d.ts
@@ -1,18 +1,33 @@
 import { Identifier } from "../interfaces";
-import { BindingCore, BindingType, ConstBinding, ConstructorBinding, FactoryBinding } from "../binder/binding";
-export interface DependencyNodeBase extends BindingCore {
+import { BindingCore, ConstBinding, ConstructorBinding, FactoryBinding } from "../binder/binding";
+import { Scope } from "../scope";
+import { DependencyGraphPlanner } from "./planner";
+export declare type ScopeDefiner = ScopedDependenencyNode | symbol;
+export interface ScopeInstance {
     /**
-     * The type of the dependency node.
+     * The ComponentCreator that defined this scope.
+     *
+     * May be SingletonSymbol as a special case for singleton-scoped services.
      */
-    type: BindingType;
+    scopeDefiner: ScopeDefiner;
+    /**
+     * Instances of scopable component creators that are in this scope,
+     * keyed by the bindingId of the binding that created the given instance.
+     */
+    instances: Map<string, ScopedDependenencyNode>;
+}
+export declare type ScopeInstanceMap = Map<Scope, ScopeInstance>;
+export interface DependencyNodeBase extends BindingCore {
     /**
      * The service identifier this node represents.
      */
     identifier: Identifier;
     /**
-     * The specific instance this node represents.
+     * The specific instance of this node relative to its containing scope.
+     * This does not correspond to the instance of the created value,
+     * as a plan may be executed several times.
      */
-    instanceId: string;
+    nodeId: string;
 }
 export interface ConstDependencyNode extends DependencyNodeBase, ConstBinding {
     type: "value";
@@ -22,10 +37,15 @@ export interface ScopedDependencyNodeBase extends DependencyNodeBase {
      * The instance of the node that defines the scope this
      * node is contained in.
      */
-    scopeOwnerInstanceId?: string;
+    scopeOwnerNodeId?: string;
 }
 export interface FactoryDependencyNode extends ScopedDependencyNodeBase, FactoryBinding {
     type: "factory";
+    /**
+     * The planner responsible for planning requests
+     * made by this factory.
+     */
+    planner: DependencyGraphPlanner;
 }
 export interface ConstructorDependencyNode extends ScopedDependencyNodeBase, ConstructorBinding {
     type: "constructor";
@@ -46,10 +66,10 @@ export interface ConstructorDependencyNode extends ScopedDependencyNodeBase, Con
      *
      * The elements may be a single DependencyNode, an array of
      * DependencyNode objects, or null.
-     * If the value is a node, the node should be resolved and rejected.
+     * If the value is a node, the node should be resolved and injected.
      * If the value is an array of nodes, the nodes should be resolved
      * and the array of resolved values injected.
-     * If the value is ```null``, then ```null``` should be injected.
+     * If the value is `null`, then `null` should be injected.
      */
     propInjectionNodes: Map<string, InjectedValue>;
 }