+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Improve error message when a class constructor binding is unresolved.
 - Only resolve relevant bindings when resolving identifiers.
 - Option to force resolve all bindings: `controller.resolveAllBindings`.
+- Support scope usage from factory / toDynamicValue bindings.
 
 # 0.7.1